+++ ChangeLog.md
@@ -49,6 +49,7 @@
 - Add ‘Edit’ menu to desktop GUI
 - Allow initialising fields of web GUI using query parameters
   (`r` for rules, `w` for words)
+- Add ‘select all’ buttons to web GUI
 
 ### Code